+++ src/main/java/com/inisw/moard/content/searchquery/SearchQueryRepository.java
@@ -1,0 +1,9 @@
+package com.inisw.moard.content.searchquery;
+
+import java.util.Optional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SearchQueryRepository extends JpaRepository<SearchQuery, Long> {
+	Optional<SearchQuery> findByQuery(String query);
+}
